Renamed directory to kenny
