+++ lib/kenny/rails_ext/rack/logger.rb
